+++ ksa_compliance/background_jobs.py
@@ -1,6 +1,11 @@
+import datetime
+from typing import Optional
+
 import frappe
 from frappe.query_builder import DocType
-import datetime
+from pypika import Order
+from pypika.queries import QueryBuilder
+
 from ksa_compliance import logger
 
 
@@ -11,22 +16,51 @@ def add_batch_to_background_queue(check_date=datetime.date.today()):
         frappe.enqueue("ksa_compliance.background_jobs.sync_e_invoices",
                        check_date=check_date,
                        queue="long",
-                       timeout=360000,
+                       timeout=3480,  # 58 minutes, so that we can run it hourly
                        job_name="Sync E-Invoices")
     except Exception as ex:
-        import traceback
-        logger.error(f"Error Occurred.{traceback.print_exc()}")
+        logger.error(f"An error occurred queueing the job", exc_info=ex)
 
 
-def sync_e_invoices(check_date=datetime.date.today()):
-    logger.info(f"Starting Job..... {check_date}")
+def sync_e_invoices(check_date: Optional[datetime.datetime | datetime.date] = None, batch_size: int = 100,
+                    dry_run: bool = False):
+    prefix = '[Dry run] ' if dry_run else ''
+    logger.info(f"{prefix}Syncing with ZATCA in batches of {batch_size}")
+    if check_date:
+        logger.info(f"{prefix}Limiting sync to >= date: {check_date}")
+
+    offset = 0
+    while True:
+        query = build_query(check_date, offset, batch_size)
+        additional_field_docs = query.run(as_dict=True)
+        if not additional_field_docs:
+            break
+
+        logger.info(f"{prefix}Syncing {len(additional_field_docs)} at offset {offset}")
+        offset += len(additional_field_docs)
+
+        for doc in additional_field_docs:
+            try:
+                logger.info(f"{prefix}Submitting {doc.name}")
+                if dry_run:
+                    continue
+
+                adf_doc = frappe.get_doc("Sales Invoice Additional Fields", doc.name)
+                adf_doc.submit()
+                frappe.db.commit()
+            except Exception as e:
+                # Review: Should we roll back?
+                logger.error(f"{prefix}Error submitting {doc.name}", exc_info=e)
+
+    logger.info(f"{prefix}Sync Done")
+
+
+def build_query(check_date: datetime.datetime | datetime.date, offset: int, limit: int) -> QueryBuilder:
     batch_status = ["Ready For Batch", "Resend", "Corrected"]
-    # TODO: Revisit the where clause
     doctype = DocType("Sales Invoice Additional Fields")
     query = (frappe.qb.from_(doctype).select(doctype.name)
-    .where(
-        (doctype.integration_status.isin(batch_status)) & (doctype.creation > check_date) & (doctype.docstatus == 0)))
-    additional_field_docs = query.run(as_dict=True)
-    for doc in additional_field_docs:
-        adf_doc = frappe.get_doc("Sales Invoice Additional Fields", doc.name)
-        adf_doc.submit()
+             .where((doctype.integration_status.isin(batch_status)) & (doctype.docstatus == 0)))
+    if check_date:
+        query = query.where(doctype.creation > check_date)
+    query = query.orderby(doctype.creation, order=Order.asc).offset(offset).limit(limit)
+    return query